refactor(core): extract budget compat model to separate file
