+++ addons/smart_construction_core/models/cost_domain.py
@@ -1,4 +1,10 @@
 # -*- coding: utf-8 -*-
+"""
+成本域聚合工具 / 领域服务。
+
+兼容模型 project.budget.line 已移动至 budget_compat.py，
+避免 cost_domain.py 里同时承担“领域服务 + 兼容层”的职责。
+"""
 from odoo import api, fields, models
 from odoo.exceptions import UserError
 
@@ -436,15 +442,3 @@ class ProjectProgressEntry(models.Model):
                 blocked_states=("paused", "closed", "closing"),
             )
         return super().create(vals_list)
-
-
-class ProjectBudgetLineCompat(models.Model):
-    """
-    兼容层：历史模型 project.budget.line -> 现用 project.budget.boq.line。
-    避免数据库残留记录升级时报“KeyError: 'project.budget.line'”。
-    """
-
-    _name = "project.budget.line"
-    _description = "项目预算行(兼容层)"
-    _inherit = "project.budget.boq.line"
-    _table = "project_budget_boq_line"
